Fix qmath(3) build problems with GCC 8.
Discussed with:	asomers
Sponsored by:	Klara Systems
Differential Revision:	https://reviews.freebsd.org/D21442


git-svn-id: svn+ssh://svn.freebsd.org/base/head@351596 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/sys/qmath.h
@@ -267,7 +267,7 @@ typedef	u64q_t		umaxq_t;
 
 /* Left shift an integral value to align with the int bits of 'q'. */
 #define	Q_SHL(q, iv) \
-    (Q_LTZ(iv) ? -(Q_ABS(iv) << Q_NFBITS(q)) :	\
+    (Q_LTZ(iv) ? -(int64_t)(Q_ABS(iv) << Q_NFBITS(q)) :	\
     Q_TC(q, iv) << Q_NFBITS(q))
 
 /* Calculate the relative fractional precision between 'a' and 'b' in bits. */
